fix styles &  structure
