+++ src/app/App.tsx
@@ -1,7 +1,6 @@
 import { useRoutes } from 'react-router-dom';
 import routes from './routes';
 import { Suspense } from 'react';
-import './styles/app.scss';
 
 function App () {
     const elements = useRoutes(routes);
