--- src/app/styles/app.scss
@@ -1,117 +1,0 @@
-.app {
-    &__container {
-        max-width: 1024px;
-        margin: 0 auto;
-    }
-    &__wrapper {
-        display: flex;
-        width: 100%;
-        flex-direction: row;
-        flex: 0 1 auto;
-        height: 100%;
-        background-color: rgb(236, 236, 236);
-        overflow-y: auto;
-    }
-
-    &__header {
-        width: 100%;
-        background-color: rgba(255, 255, 255, 100);
-        width: 100%;
-        padding: 36px 0;
-
-        &_grid {
-            display: grid;
-            justify-content: center;
-            align-items: center;
-        }
-    }
-
-    &__body {
-        font-size: 20px;
-
-        flex: 1 1 auto;
-        height: 100%;
-        background-color: rgb(236, 236, 236);
-        padding-bottom: 20px;
-    }
-    &__title {
-        height: 5rem;
-        font-size: 36px;
-        font-weight: 500;
-    }
-
-    &__ribbon {
-        font-size: 17px;
-        font-weight: 500;
-        display: flex;
-    }
-}
-
-.content {
-    &__container {
-        min-height: 100%;
-        max-width: 1024px;
-        width: 100%;
-        margin: 0 auto;
-        padding: 24px 64px 20px 64px;
-    }
-
-    &__body {
-        flex: 1 1 auto;
-        height: 100%;
-        background-color: rgb(236, 236, 236);
-        padding-bottom: 20px;
-    }
-
-    &__button {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        box-sizing: border-box;
-        padding: 10px;
-        border-bottom: 2px solid rgb(255, 255, 255);
-        color: black;
-
-        &.active {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            box-sizing: border-box;
-            color: var(--primary-color);
-            padding: 10px;
-            border-bottom: 2px solid var(--primary-color);
-        }
-    }
-}
-
-@media (max-width: 1600px) {
-    .app {
-        &__content-head {
-            background-image: none;
-            background-color: rgba(255, 255, 255, 100);
-        }
-    }
-}
-
-@media (max-width: 767px) {
-    .app {
-        &__content-head {
-            background: none;
-            background-color: rgba(255, 255, 255, 100);
-        }
-
-        &__content-container {
-            flex: 0 1 auto;
-            flex-direction: column;
-            justify-content: center;
-            align-items: center;
-            min-height: 100%;
-            width: 100%;
-            margin: 0 auto;
-            padding: 0 4px;
-        }
-    }
-    .app__body {
-        width: 100%;
-    }
-}
