+++ src/entities/inventoryDocs/components/inventoryContent/inventoryContent.scss
@@ -1,0 +1,22 @@
+// .products-list {
+//     margin-top: 20px;
+//     align-items: center;
+//     justify-content: center;
+//     font-size: 14px;
+
+//     &__total {
+//         display: flex;
+//         line-height: 24px;
+//         min-height: 48px;
+//         background-color: #fff;
+//         margin-top: 3px;
+//         align-items: center;
+//         font-weight: bold;
+//         justify-content: flex-end;
+//         padding: 0 12px;
+//     }
+
+//     &__sum {
+//         margin-left: 20px;
+//     }   
+// }
