--- src/entities/inventoryDocs/components/inventoryDocs.scss
@@ -1,9 +1,0 @@
-.inventory {
-    &__item {
-        display: flex;
-        margin-top: 16px;
-        font-size: 16px;
-        background-color: #fff;
-        padding: 16px 36px;
-    }
-}
