--- src/entities/inventoryDocs/components/inventoryList.scss
@@ -1,98 +1,0 @@
-.products-list {
-    margin-top: 20px;
-    align-items: center;
-    justify-content: center;
-    font-size: 14px;
-    &__header {
-        display: flex;
-        background-color: #fff;
-        line-height: 24px;
-        min-height: 48px;
-        justify-content: flex-end;
-        align-items: flex-end;
-        font-weight: bold;
-    }
-
-    &__priceColumn {
-        display: flex;
-        width: 100px;
-        flex: 0 0 auto;
-        padding: 0 12px;
-        justify-content: flex-end;
-    }
-
-    &__nameColumn {
-        flex: 1 1 auto;
-        padding: 0 12px;
-    }
-
-    &__total {
-        display: flex;
-        line-height: 24px;
-        min-height: 48px;
-        background-color: #fff;
-        margin-top: 3px;
-        align-items: center;
-        font-weight: bold;
-        justify-content: flex-end;
-        padding: 0 12px;
-    }
-    &__sum {
-        margin-left: 20px;
-    }
-}
-
-.createXLSX {
-    display: block;
-    padding: 5px 15px;
-    background-color: #fff;
-    font-size: 16px;
-    border: 1px solid black;
-    border-radius: 10px;
-    margin: 10px 0;
-}
-
-.product-card {
-    display: flex;
-    line-height: 24px;
-    min-height: 48px;
-    background-color: #fff;
-    margin-top: 3px;
-    align-items: center;
-    cursor: pointer;
-    &__price,
-    &__quantity,
-    &__sum,
-    &__remove {
-        display: flex;
-        justify-content: flex-end;
-        padding: 0 12px;
-        width: 100px;
-        flex: 0 0 auto;
-    }
-
-    &__remove {
-        height: 100%;
-        width: 30px;
-        color: red;
-        font-weight: bold;
-        font-size: 24px;
-    }
-    &__remove:hover {
-        -webkit-text-stroke-width: 2px;
-        -webkit-text-stroke-color: white;
-    }
-
-    &__number {
-        display: flex;
-        justify-content: flex-end;
-        padding: 0 12px;
-        width: 50px;
-        flex: 0 0 auto;
-    }
-
-    &__title {
-        flex: 1 1 auto;
-        padding: 0 12px;
-    }
-}
