+++ src/entities/inventoryDocs/components/inventoryListItem/inventoryDocs.scss
@@ -1,0 +1,7 @@
+.inventory__item {
+    display: flex;
+    margin-top: 16px;
+    font-size: 16px;
+    background-color: #fff;
+    padding: 16px 36px;
+}
