+++ src/entities/inventoryDocs/components/inventoryProductCard/inventoryProductCard.scss
@@ -1,0 +1,45 @@
+.inventory-card {
+    display: flex;
+    line-height: 24px;
+    min-height: 48px;
+    background-color: #fff;
+    margin-top: 3px;
+    align-items: center;
+    cursor: pointer;
+
+    &__price,
+    &__quantity,
+    &__sum,
+    &__remove {
+        display: flex;
+        justify-content: flex-end;
+        padding: 0 12px;
+        width: 100px;
+        flex: 0 0 auto;
+    }
+
+    &__remove {
+        height: 100%;
+        width: 30px;
+        color: red;
+        font-weight: bold;
+        font-size: 24px;
+    }
+    &__remove:hover {
+        -webkit-text-stroke-width: 2px;
+        -webkit-text-stroke-color: white;
+    }
+
+    &__number {
+        display: flex;
+        justify-content: flex-end;
+        padding: 0 12px;
+        width: 50px;
+        flex: 0 0 auto;
+    }
+
+    &__title {
+        flex: 1 1 auto;
+        padding: 0 12px;
+    }
+}
