+++ src/entities/inventoryDocs/index.ts
@@ -1,4 +1,4 @@
 import { inventoryDocsAPI } from './reducers/inventoryDocsService';
-import InventoryList from './components/inventoryList';
-import InventoryContent from './components/inventoryContent';
+import InventoryList from './components/inventoryList/inventoryList';
+import InventoryContent from './components/inventoryContent/inventoryContent';
 export { inventoryDocsAPI, InventoryContent, InventoryList };
