+++ src/entities/products/components/categoryCard/categoryStyles.scss
