+++ src/entities/products/index.ts
@@ -1,7 +1,7 @@
 import { productsAPI } from '@entities/products/model/productsService';
-import ProductCard from './components/productCard';
+import ProductCard from './components/productCard/productCard';
 import ProductList from './components/productList';
-import CategoryCard from './components/categoryCard';
+import CategoryCard from './components/categoryCard/categoryCard';
 import CategoriesList from './components/categoriesList/categoriesList';
 
 export { ProductCard, ProductList, CategoryCard, CategoriesList, productsAPI };
