+++ src/pages/inventoryPage/components/inventoryPage.async.tsx
