+++ src/pages/inventoryPage/index.ts
