+++ src/pages/profilePage/components/profile.scss
