+++ src/shared/layouts/appBody/appBody.scss
@@ -1,0 +1,22 @@
+.app__body {
+    font-size: 20px;
+
+    flex: 1 1 auto;
+    height: 100%;
+    background-color: rgb(236, 236, 236);
+    padding-bottom: 20px;
+}
+
+.content__container {
+    min-height: 100%;
+    max-width: 1024px;
+    width: 100%;
+    margin: 0 auto;
+    padding: 24px 64px 20px 64px;
+}
+
+@media (max-width: 767px) {
+    .app__body {
+        width: 100%;
+    }
+}
