+++ src/shared/layouts/appHeader/appHeader.tsx
@@ -1,4 +1,5 @@
 import { type ReactNode } from 'react';
+import './appHeader.scss';
 
 interface AppHeaderProps {
     children?: ReactNode
