+++ src/shared/layouts/appLayout/appLayout.scss
@@ -1,0 +1,21 @@
+.app {
+    &__wrapper {
+        display: flex;
+        width: 100%;
+        flex-direction: row;
+        flex: 0 1 auto;
+        height: 100%;
+        background-color: rgb(236, 236, 236);
+        overflow-y: auto;
+    }
+
+    &__content {
+        display: flex;
+        width: 100%;
+        flex-direction: column;
+        flex: 0 1 auto;
+        height: 100%;
+        background-color: rgb(236, 236, 236);
+        overflow-y: auto;
+    }
+}
