+++ src/shared/layouts/appLayout/appLayout.tsx
@@ -1,4 +1,5 @@
 import { type ReactNode } from 'react';
+import './appLayout.scss';
 
 interface LayoutProps {
     children: ReactNode
