+++ src/shared/ui/calendar/calendar.scss
@@ -11,19 +11,6 @@
     align-items: flex-end;
 }
 
-.cut-short {
-    width: 50px;
-}
-.placeholder {
-    color: #65657b;
-    left: 20px;
-    line-height: 14px;
-    pointer-events: none;
-    position: absolute;
-    transform-origin: 0 50%;
-    transition: transform 200ms, color 200ms;
-    top: 20px;
-}
 .date-placeholder {
     background-color: inherit;
     border-radius: 10px;
@@ -31,30 +18,7 @@
     left: 20px;
     position: absolute;
     top: -5px;
-    width: 100px;
+    width: 100%;
     font-size: 13.5px;
     color: #808097;
-}
-
-.datepicker:focus ~ .placeholder,
-.datepicker:not(:placeholder-shown) ~ .placeholder {
-    transform: translateY(-25px) translateX(10px) scale(0.75);
-}
-
-.datepicker:not(:placeholder-shown) ~ .placeholder {
-    color: #808097;
-}
-
-.datepicker:focus ~ .placeholder {
-    color: #808097;
-}
-
-
-.cut-short {
-    width: 50px;
-}
-
-.datepicker:focus ~ .cut,
-.datepicker:not(:placeholder-shown) ~ .cut {
-    transform: translateY(8px);
 }
