+++ src/widgets/documentPageBody/components/distributionContent/distribution.tsx
