+++ src/widgets/documentPageBody/components/orderingContent/ordering.scss
@@ -1,0 +1,8 @@
+.content {
+    &__body {
+        flex: 1 1 auto;
+        height: 100%;
+        background-color: rgb(236, 236, 236);
+        padding-bottom: 20px;
+    }
+}
