+++ src/widgets/documentPageBody/consts/docsNavLinks.ts
