+++ src/widgets/documentPageBody/consts/tableHeaders.ts
