+++ src/widgets/documentPageBody/consts/types.ts
