+++ src/widgets/documentPageBody/index.ts
@@ -1,5 +1,5 @@
-import Distribution from './components/distribution';
-import Order from './components/order';
-import Ordering from './components/ordering';
+import Distribution from './components/distributionContent/distribution';
+import Order from './components/orderContent/order';
+import Ordering from './components/orderingContent/ordering';
 
 export { Order, Ordering, Distribution };
