+++ src/widgets/navBar/components/navBar/navBar.scss
@@ -11,15 +11,7 @@
     overflow-x: hidden;
     background-color: rgb(255, 255, 255);
 }
-.app__content {
-    display: flex;
-    width: 100%;
-    flex-direction: column;
-    flex: 0 1 auto;
-    height: 100%;
-    background-color: rgb(236, 236, 236);
-    overflow-y: auto;
-}
+
 .navigation {
     height: 100%;
 }
